fix up DI server to use governator
