+++ eureka-server-governator/build.gradle
@@ -3,22 +3,15 @@ apply plugin: 'jetty'
 
 dependencies {
     compile project(':eureka-core')
-    compile project(':eureka-client-archaius2')
-    runtime 'xerces:xercesImpl:2.4.0'
+    compile project(':eureka-client')
     compile "com.netflix.governator:governator:${governatorVersion}"
     compile "com.netflix.governator:governator-servlet:${governatorVersion}"
-    compile "com.netflix.karyon:karyon3-core:${karyon3Version}"
-    compile "com.netflix.karyon:karyon3-archaius2:${karyon3Version}"
-    compile "com.netflix.karyon:karyon3-jetty:${karyon3Version}"
     compile "com.sun.jersey:jersey-server:${jerseyVersion}"
     compile "com.sun.jersey:jersey-servlet:${jerseyVersion}"
     compile "com.sun.jersey.contribs:jersey-guice:${jerseyVersion}"
     compile 'org.slf4j:slf4j-log4j12:1.6.1'
     runtime 'org.codehaus.jettison:jettison:1.2' 
     providedCompile "javax.servlet:servlet-api:${servletVersion}"
-
-    // for now to keep governator and karyon3 in sync until karyon3 stabilize
-    compile "com.netflix.governator:governator-jetty:${governatorVersion}"
 
     testCompile project(':eureka-test-utils')
     testCompile 'org.mockito:mockito-core:1.10.19'
