+++ eureka-server-governator/src/main/java/com/netflix/eureka/EurekaContextListener.java
@@ -45,7 +45,7 @@ public class EurekaContextListener extends GovernatorServletContextListener {
         JsonXStream.getInstance().registerConverter(new V1AwareInstanceInfoConverter(), XStream.PRIORITY_VERY_HIGH);
         XmlXStream.getInstance().registerConverter(new V1AwareInstanceInfoConverter(), XStream.PRIORITY_VERY_HIGH);
 
-        LifecycleInjector injector = EurekaInjectorCreator.createInjector(false);
+        LifecycleInjector injector = EurekaInjectorCreator.createInjector();
         serverContext = injector.getInstance(EurekaServerContext.class);
         return injector;
     }
