+++ eureka-server-governator/src/main/java/com/netflix/eureka/guice/Ec2EurekaServerModule.java
@@ -6,31 +6,31 @@ import com.netflix.eureka.DefaultEurekaServerConfig;
 import com.netflix.eureka.DefaultEurekaServerContext;
 import com.netflix.eureka.EurekaServerConfig;
 import com.netflix.eureka.EurekaServerContext;
+import com.netflix.eureka.aws.AwsBinderDelegate;
 import com.netflix.eureka.cluster.PeerEurekaNodes;
 import com.netflix.eureka.registry.AbstractInstanceRegistry;
+import com.netflix.eureka.registry.AwsInstanceRegistry;
 import com.netflix.eureka.registry.InstanceRegistry;
 import com.netflix.eureka.registry.PeerAwareInstanceRegistry;
-import com.netflix.eureka.registry.PeerAwareInstanceRegistryImpl;
 import com.netflix.eureka.resources.DefaultServerCodecs;
 import com.netflix.eureka.resources.ServerCodecs;
-import com.netflix.karyon.conditional.ConditionalOnLocalDev;
 
 /**
  * @author David Liu
  */
-@ConditionalOnLocalDev
-public class LocalDevEurekaServerModule extends AbstractModule {
+public class Ec2EurekaServerModule extends AbstractModule {
     @Override
     protected void configure() {
-        // server bindings
         bind(EurekaServerConfig.class).to(DefaultEurekaServerConfig.class).in(Scopes.SINGLETON);
         bind(PeerEurekaNodes.class).in(Scopes.SINGLETON);
 
+        bind(AwsBinderDelegate.class).asEagerSingleton();
+
         // registry and interfaces
-        bind(PeerAwareInstanceRegistryImpl.class).asEagerSingleton();
-        bind(InstanceRegistry.class).to(PeerAwareInstanceRegistryImpl.class);
-        bind(AbstractInstanceRegistry.class).to(PeerAwareInstanceRegistryImpl.class);
-        bind(PeerAwareInstanceRegistry.class).to(PeerAwareInstanceRegistryImpl.class);
+        bind(AwsInstanceRegistry.class).asEagerSingleton();
+        bind(InstanceRegistry.class).to(AwsInstanceRegistry.class);
+        bind(AbstractInstanceRegistry.class).to(AwsInstanceRegistry.class);
+        bind(PeerAwareInstanceRegistry.class).to(AwsInstanceRegistry.class);
 
         bind(ServerCodecs.class).to(DefaultServerCodecs.class).in(Scopes.SINGLETON);
 
@@ -39,11 +39,11 @@ public class LocalDevEurekaServerModule extends AbstractModule {
 
     @Override
     public boolean equals(Object obj) {
-        return LocalDevEurekaServerModule.class.equals(obj.getClass());
+        return Ec2EurekaServerModule.class.equals(obj.getClass());
     }
 
     @Override
     public int hashCode() {
-        return LocalDevEurekaServerModule.class.hashCode();
+        return Ec2EurekaServerModule.class.hashCode();
     }
 }
