+++ eureka-server-governator/src/main/resources/eureka-client.properties
