+++ eureka-server-governator/src/main/resources/eureka-server.properties
