+++ eureka-server-governator/src/main/resources/log4j.properties
@@ -1,6 +1,5 @@
 log4j.rootCategory=INFO,stdout
 log4j.appender.stdout=org.apache.log4j.ConsoleAppender
-#log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
-log4j.appender.stdout.layout=com.netflix.logging.log4jAdapter.NFPatternLayout
+log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
 log4j.appender.stdout.layout.ConversionPattern=%d %-5p %C:%L [%t] [%M] %m%n
 log4j.logger.asyncAppenders=INFO,stdout
