+++ settings.gradle
@@ -1,2 +1,10 @@
 rootProject.name='eureka' 
-include 'eureka-client', 'eureka-client-jersey2', 'eureka-client-archaius2','eureka-server','eureka-server-karyon3','eureka-core','eureka-resources','eureka-examples', 'eureka-test-utils'
+include 'eureka-client',
+        'eureka-client-jersey2',
+        'eureka-client-archaius2',
+        'eureka-server',
+        'eureka-server-governator',
+        'eureka-core',
+        'eureka-resources',
+        'eureka-examples',
+        'eureka-test-utils'
